# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Expense {
  id: number;
  amount: number;
  category: string;
  description: string;
  date: string;
  createdAt: string;
}

export interface CreateExpenseBody {
  amount: number;
  category: string;
  description: string;
  date: string;
}

export interface CategoryTotal {
  category: string;
  total: number;
  count: number;
}

export interface MonthlyTotal {
  month: string;
  total: number;
  count: number;
}

export interface ExpenseSummary {
  totalSpending: number;
  categoryBreakdown: CategoryTotal[];
  monthlyTotals: MonthlyTotal[];
  /** @nullable */
  topCategory: string | null;
  /** @nullable */
  insight: string | null;
  /** @nullable */
  budgetLimit: number | null;
  budgetExceeded: boolean;
  /** @nullable */
  remainingBudget: number | null;
}

export interface Budget {
  /** @nullable */
  limit: number | null;
}

export interface SetBudgetBody {
  limit: number;
}

export type ListExpensesParams = {
  /**
   * Filter by category
   */
  category?: string;
  /**
   * Filter by month (YYYY-MM)
   */
  month?: string;
};

export type GetExpenseSummaryParams = {
  /**
   * Filter summary to a specific month (YYYY-MM)
   */
  month?: string;
};