# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all expenses, optionally filtered by category
 * @summary List expenses
 */
export const ListExpensesQueryParams = zod.object({
  category: zod.coerce.string().optional().describe("Filter by category"),
  month: zod.coerce.string().optional().describe("Filter by month (YYYY-MM)"),
});

export const ListExpensesResponseItem = zod.object({
  id: zod.number(),
  amount: zod.number(),
  category: zod.string(),
  description: zod.string(),
  date: zod.string(),
  createdAt: zod.string(),
});
export const ListExpensesResponse = zod.array(ListExpensesResponseItem);

/**
 * Create a new expense
 * @summary Add expense
 */
export const AddExpenseBody = zod.object({
  amount: zod.number(),
  category: zod.string(),
  description: zod.string(),
  date: zod.string(),
});

/**
 * Delete an expense by ID
 * @summary Delete expense
 */
export const DeleteExpenseParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * Returns total spending, monthly totals, category breakdown, and smart insights
 * @summary Get expense summary
 */
export const GetExpenseSummaryQueryParams = zod.object({
  month: zod.coerce
    .string()
    .optional()
    .describe("Filter summary to a specific month (YYYY-MM)"),
});

export const GetExpenseSummaryResponse = zod.object({
  totalSpending: zod.number(),
  categoryBreakdown: zod.array(
    zod.object({
      category: zod.string(),
      total: zod.number(),
      count: zod.number(),
    }),
  ),
  monthlyTotals: zod.array(
    zod.object({
      month: zod.string(),
      total: zod.number(),
      count: zod.number(),
    }),
  ),
  topCategory: zod.string().nullable(),
  insight: zod.string().nullable(),
  budgetLimit: zod.number().nullable(),
  budgetExceeded: zod.boolean(),
  remainingBudget: zod.number().nullable(),
});

/**
 * Returns the current budget limit
 * @summary Get budget
 */
export const GetBudgetResponse = zod.object({
  limit: zod.number().nullable(),
});

/**
 * Set monthly budget limit
 * @summary Set budget
 */
export const SetBudgetBody = zod.object({
  limit: zod.number(),
});

export const SetBudgetResponse = zod.object({
  limit: zod.number().nullable(),
});