# lib/api-zod/src/generated/types/expenseSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryTotal } from "./categoryTotal";
import type { MonthlyTotal } from "./monthlyTotal";

export interface ExpenseSummary {
  totalSpending: number;
  categoryBreakdown: CategoryTotal[];
  monthlyTotals: MonthlyTotal[];
  /** @nullable */
  topCategory: string | null;
  /** @nullable */
  insight: string | null;
  /** @nullable */
  budgetLimit: number | null;
  budgetExceeded: boolean;
  /** @nullable */
  remainingBudget: number | null;
}