# lib/api-zod/src/generated/types/monthlyTotal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyTotal {
  month: string;
  total: number;
  count: number;
}